Remove kroger.com from disposable email domains list (#308)
* Remove kroger.com from disposable email domains list

kroger.com is a legitimate corporate domain (Fortune 500 company) and should
not be flagged as a disposable email service. This change allows legitimate
corporate users with @kroger.com email addresses to sign up for services
using this gem.

* Also remove krogerco.com from disposable email domains list

krogerco.com is another legitimate corporate domain (Kroger Company) and should
not be flagged as a disposable email service. This change allows legitimate
corporate users with @krogerco.com email addresses to sign up for services
using this gem.

* Add krogerco.com back to disposable domains list

krogerco.com is not a valid corporate domain and should remain in the
disposable email domains list. Only kroger.com should be removed as it
is a legitimate Fortune 500 corporate domain.